+++ skv/t_kv.go
@@ -16,8 +16,8 @@ package skv
 
 type KvInterface interface {
 	KvGet(key []byte) *Reply
-	KvPut(key, value []byte, ttl uint32) *Reply
-	KvPutJson(key []byte, value interface{}, ttl uint32) *Reply
+	KvPut(key, value []byte, ttl int64) *Reply
+	KvPutJson(key []byte, value interface{}, ttl int64) *Reply
 	KvDel(keys ...[]byte) *Reply
 	KvScan(cursor, end []byte, limit uint32) *Reply
 	KvIncrby(key []byte, step int64) *Reply
